Make geometry tests C++
